reorganized component files
