+++ src/app/free/free-routing.module.ts
@@ -1,9 +1,9 @@
 import { RouterModule, Routes } from '@angular/router';
 
 import { NgModule } from '@angular/core';
-import { WorkoutsCalendarComponent } from './workouts-calendar/workouts-calendar.component';
+import { WorkoutsCalendarComponent } from './workouts/workouts-calendar/workouts-calendar.component';
 import { WorkoutsComponent } from './workouts/workouts.component';
-import { WorkoutsListComponent } from './workouts-list/workouts-list.component';
+import { WorkoutsListComponent } from './workouts/workouts-list/workouts-list.component';
 
 const routes: Routes = [
   { path: '', redirectTo: '/free/workouts/calendar', pathMatch: 'full' },
