+++ src/app/pro/cycles/edit-cycle/delete-cycle/delete-cycle.component.ts
