+++ src/app/pro/pro.module.ts
@@ -5,7 +5,7 @@ import { NgModule } from '@angular/core';
 import { ProRoutingModule } from './pro-routing.module';
 import { EditCycleComponent } from './cycles/edit-cycle/edit-cycle.component';
 import { ReactiveFormsModule } from '@angular/forms';
-import { DeleteCycleComponent } from './cycles/delete-cycle/delete-cycle.component';
+import { DeleteCycleComponent } from './cycles/edit-cycle/delete-cycle/delete-cycle.component';
 
 @NgModule({
   declarations: [CyclesComponent, EditCycleComponent, DeleteCycleComponent],
